svn merge -r 26775:27023 https://dart.googlecode.com/svn/branches/bleeding_edge trunk

git-svn-id: http://dart.googlecode.com/svn/trunk@27025 260f80e4-7a28-3924-810f-c04153c831b5

+++ tests/html/wheelevent_test.dart
@@ -19,7 +19,7 @@ main() {
     element.onMouseWheel.listen(expectAsync1((e) {
       expect(e.screen.x, 100);
       expect(e.deltaX, 0);
-      expect(e.deltaY, 240);
+      expect(e.deltaY.toDouble(), 240.0);
       expect(e.deltaMode, isNotNull);
     }));
     var event = new WheelEvent(eventType,
@@ -36,7 +36,7 @@ main() {
     element.onMouseWheel.listen(expectAsync1((e) {
       expect(e.screen.x, 100);
       expect(e.deltaX, 0);
-      expect(e.deltaY, 240);
+      expect(e.deltaY.toDouble(), 240.0);
     }));
     var event = new WheelEvent(eventType,
       deltaX: 0,
